server/requirements.txt
edge-tts
asyncio
pdfminer.six
python-docx
